# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BananaPong",
    "description": "This is the documentation of the BananaPong server. Remember that the JWT is necessary for all the operations (except the ones in the auth tag), Make sure you input a correct JWT token in the Authorize menu.",
    "version": "2.0",
    "contact": {
      "email": "droge@student.42.fr"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 65000
          },
          "log": {
            "type": "string",
            "example": "droge"
          },
          "username": {
            "type": "string",
            "example": "dimitri"
          },
          "onlineStatus": {
            "type": "boolean",
            "example": "true"
          },
          "profileCompleted": {
            "type": "boolean",
            "example": "true"
          },
          "rank": {
            "type": "number",
            "example": 100
          },
          "friends": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "example": [
              65255,
              65350
            ]
          },
          "blocked": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "example": [
              65243,
              65379
            ]
          },
          "otp": {
            "type": "string",
            "example": "lnrtuvcbdefghijk"
          },
          "avatarId": {
            "type": "number",
            "example": 65000
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "JWT Token is invalid or empty"
      },
      "OK": {
        "description": "Expected response"
      }
    }
  },
  "paths": {
...
    "/user": {
      "get": {
        "tags": [
          "user id"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Returns all the users of the server",
        "operationId": "IdController_getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      },
      "post": {
        "tags": [
          "user id"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Creates a user object (only id and log will be saved on the database)",
        "operationId": "IdController_createUser",
        "parameters": [],
        "requestBody": {
          "description": "New user object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "log": {
                    "type": "string",
                    "example": "droge"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/OK"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "user id"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Returns the user object (specified by id)",
        "operationId": "IdController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "User doesn't exist in database"
          }
        }
      },
      "put": {
        "tags": [
          "user id"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Modifies the user object (specified by id)",
        "operationId": "IdController_updateUser",
        "parameters": [
          {
            "name": "id",
            "description": "Id of the user",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "Updated user object (id and avatarId cannot be modified)",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "User doesn't exist in database"
          }
        }
      },
      "delete": {
        "tags": [
          "user id"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete the user object (specified by id)",
        "operationId": "IdController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "User doesn't exist in database"
          }
        }
      }
    },
    "/user/{id}/avatar": {
      "get": {
        "tags": [
          "user id"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Returns the user's avatar (specified by id)",
        "operationId": "IdController_getAvatar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "User doesn't exist in database"
          }
        }
      },
      "put": {
        "tags": [
          "user id"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Returns the user's avatar (specified by id)",
        "operationId": "IdController_setAvatar",
        "parameters": [
          {
            "name": "id",
            "description": "Id of the user",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "Updated avatar",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "description": "new Avatar",
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    }
  }
}